Requisições de patch e delete de category e brand

+++ api/routers/brand.py
@@ -1,10 +1,9 @@
-from fastapi import APIRouter, HTTPException
-from models.models import new_category
+from fastapi import HTTPException, APIRouter, UploadFile, File, Form, Depends
+from dependencies import token
 from database.connection import mysql_connection
 import uuid
 import time
 import os
-from fastapi import HTTPException, APIRouter, UploadFile, File, Form
 
 router = APIRouter()
 
@@ -31,11 +30,12 @@ async def get_brands():
 
 
 @router.post("/brand", tags=['Marca'])
-async def create_brand(name: str = Form(), brand_logo: UploadFile = File(None), brand_logo_black: UploadFile = File(None)):
+async def create_brand(name: str = Form(), brand_logo: UploadFile = File(None), brand_logo_black: UploadFile = File(None), current_user: int = Depends(token.is_admin)):
     try:
         cursor = mysql_connection.cursor(dictionary=True)
         id = int.from_bytes(
             uuid.uuid4().bytes[:4], byteorder="big") % (2 ** 30)
+
         if brand_logo is not None:
             filename_brand_logo = str(time.time()) + \
                 brand_logo.filename.replace(' ', '_')
@@ -43,6 +43,7 @@ async def create_brand(name: str = Form(), brand_logo: UploadFile = File(None), 
             with open(file_path, "wb") as f:
                 f.write(brand_logo.file.read())
             print(filename_brand_logo)
+
         if brand_logo_black is not None:
             filename_brand_logo_black = str(
                 time.time()) + brand_logo_black.filename.replace(' ', '_')
@@ -57,16 +58,19 @@ async def create_brand(name: str = Form(), brand_logo: UploadFile = File(None), 
                 (id, name, str(filename_brand_logo),
                  str(filename_brand_logo_black))
             )
+
         elif brand_logo is None and brand_logo_black is not None:
             cursor.execute(
                 "INSERT INTO BRAND (id, name, brand_logo_black) VALUES (%s, %s, %s)",
-                (id, name, str('public/brand/' + filename_brand_logo_black))
+                (id, name, str(filename_brand_logo_black))
             )
+
         elif brand_logo is not None and brand_logo_black is None:
             cursor.execute(
                 "INSERT INTO BRAND (id, name, brand_logo) VALUES (%s, %s, %s)",
-                (id, name, str('public/brand/' + filename_brand_logo))
+                (id, name, str(filename_brand_logo))
             )
+
         else:
             cursor.execute(
                 "INSERT INTO BRAND (id, name) VALUES (%s, %s)",
@@ -90,11 +94,11 @@ async def create_brand(name: str = Form(), brand_logo: UploadFile = File(None), 
 
 
 @router.patch("/brand", tags=["Marca"])
-async def update_brand(old_name: str = Form(), name: str = Form(None), brand_logo: UploadFile = File(None), brand_logo_black: UploadFile = File(None)):
+async def update_brand(id: str = Form(), name: str = Form(None), brand_logo: UploadFile = File(None), brand_logo_black: UploadFile = File(None), current_user: int = Depends(token.is_admin)):
     try:
         cursor = mysql_connection.cursor(dictionary=True)
 
-        cursor.execute("SELECT * FROM BRAND WHERE NAME = %s", (old_name,))
+        cursor.execute("SELECT * FROM BRAND WHERE id = %s", (id,))
         old_brand = cursor.fetchone()
 
         new_brand = {}
@@ -102,7 +106,7 @@ async def update_brand(old_name: str = Form(), name: str = Form(None), brand_log
         if name is not None:
             new_brand["name"] = name
         else:
-            new_brand["name"] = old_name
+            new_brand["name"] = old_brand['name']
 
         if brand_logo is not None:
             filename_brand_logo = str(time.time()) + \
@@ -111,8 +115,7 @@ async def update_brand(old_name: str = Form(), name: str = Form(None), brand_log
             with open(file_path, "wb") as f:
                 f.write(brand_logo.file.read())
             print(filename_brand_logo)
-            new_brand["brand_logo"] = str(
-                'public/brand/' + filename_brand_logo)
+            new_brand["brand_logo"] = str(filename_brand_logo)
         else:
             new_brand["brand_logo"] = old_brand["brand_logo"]
 
@@ -123,15 +126,14 @@ async def update_brand(old_name: str = Form(), name: str = Form(None), brand_log
             with open(file_path, "wb") as f:
                 f.write(brand_logo_black.file.read())
             print(filename_brand_logo_black)
-            new_brand["brand_logo_black"] = str(
-                'public/brand/' + filename_brand_logo_black)
+            new_brand["brand_logo_black"] = str(filename_brand_logo_black)
         else:
             new_brand["brand_logo_black"] = old_brand["brand_logo_black"]
 
         cursor.execute(
-            "UPDATE BRAND SET name = %s, brand_logo = %s, brand_logo_black = %s WHERE name = %s",
+            "UPDATE BRAND SET name = %s, brand_logo = %s, brand_logo_black = %s WHERE id = %s",
             (new_brand["name"], new_brand["brand_logo"],
-             new_brand["brand_logo_black"], old_name)
+             new_brand["brand_logo_black"], id)
         )
         mysql_connection.commit()
 
@@ -139,16 +141,18 @@ async def update_brand(old_name: str = Form(), name: str = Form(None), brand_log
 
         # Removendo imagens antigas da api
         if brand_logo is not None:
-            filename_logo = os.path.join(old_brand["brand_logo"])
+            filename_logo = os.path.join(
+                upload_folder + '/' + old_brand["brand_logo"])
             os.remove(filename_logo)
         if brand_logo_black is not None:
-            filename_logo_black = os.path.join(old_brand["brand_logo_black"])
+            filename_logo_black = os.path.join(
+                update_brand + '/' + old_brand["brand_logo_black"])
             os.remove(filename_logo_black)
 
         cursor.execute("SELECT * FROM BRAND WHERE NAME = %s", (name,))
         brand_updated = cursor.fetchone()
         cursor.close()
-        return brand_updated
+        return True
 
     except Exception as e:
         print(e)
@@ -156,11 +160,11 @@ async def update_brand(old_name: str = Form(), name: str = Form(None), brand_log
 
 
 @router.put("/brand", tags=["Marca"])
-async def update_brand(old_name: str = Form(), name: str = Form(), brand_logo: UploadFile = File(), brand_logo_black: UploadFile = File()):
+async def update_brand(id: str = Form(), name: str = Form(), brand_logo: UploadFile = File(), brand_logo_black: UploadFile = File()):
     try:
         cursor = mysql_connection.cursor(dictionary=True)
 
-        cursor.execute("SELECT * FROM BRAND WHERE NAME = %s", (old_name,))
+        cursor.execute("SELECT * FROM BRAND WHERE id = %s", (id,))
         old_brand = cursor.fetchone()
 
         filename_brand_logo = str(time.time()) + \
@@ -177,17 +181,19 @@ async def update_brand(old_name: str = Form(), name: str = Form(), brand_logo: U
         print(filename_brand_logo_black)
 
         cursor.execute(
-            "UPDATE BRAND SET name = %s, brand_logo = %s, brand_logo_black = %s WHERE name = %s",
-            (name, filename_brand_logo, filename_brand_logo_black, old_name)
+            "UPDATE BRAND SET name = %s, brand_logo = %s, brand_logo_black = %s WHERE id = %s",
+            (name, filename_brand_logo, filename_brand_logo_black, id)
         )
         mysql_connection.commit()
 
         print("Tudo certo até aqui, iniciando a exclusão das imagens antigas")
 
         # Removendo imagens antigas da api
-        filename_logo = os.path.join(old_brand["brand_logo"])
+        filename_logo = os.path.join(
+            upload_folder + "/" + old_brand["brand_logo"])
         os.remove(filename_logo)
-        filename_logo_black = os.path.join(old_brand["brand_logo_black"])
+        filename_logo_black = os.path.join(
+            upload_folder + "/" + old_brand["brand_logo_black"])
         os.remove(filename_logo_black)
         mysql_connection.commit()
 
@@ -201,20 +207,24 @@ async def update_brand(old_name: str = Form(), name: str = Form(), brand_logo: U
         return e
 
 
-@router.delete("/brand", tags=["Marca"])
-async def delete_brand(name: str = Form()):
+@router.delete("/brand/{id}", tags=["Marca"])
+async def delete_brand(id: str, current_user: int = Depends(token.is_admin)):
     try:
         cursor = mysql_connection.cursor(dictionary=True)
-        cursor.execute("SELECT * FROM BRAND WHERE NAME = %s", (name,))
+        cursor.execute("SELECT * FROM BRAND WHERE id = %s", (id,))
         data = cursor.fetchone()
 
-        cursor.execute("DELETE FROM BRAND WHERE NAME LIKE %s", (name,))
+        cursor.execute("DELETE FROM BRAND WHERE id LIKE %s", (id,))
 
-        filename_logo = os.path.join(data["brand_logo"])
+        if data['brand_logo'] is not None:
+            filename_logo = os.path.join(
+                upload_folder + "/" + data["brand_logo"])
+            os.remove(filename_logo)
 
-        os.remove(filename_logo)
-        filename_logo_black = os.path.join(data["brand_logo_black"])
-        os.remove(filename_logo_black)
+        if data['brand_logo_black'] is not None:
+            filename_logo_black = os.path.join(
+                upload_folder + "/" + data["brand_logo_black"])
+            os.remove(filename_logo_black)
         mysql_connection.commit()
 
         # # Fechar o cursor e retornar o ID do produto
